change velocity on collision

+++ zoe/src/zoe/display/Camera.cpp
@@ -53,7 +53,7 @@ Camera::Camera() = default;
 
 //Camera2D
 
-Camera2D::Camera2D(const vec2& position, const float& width, const float& height) : width(width), height(height) {
+Camera2D::Camera2D(const vec2& position, const double& width, const double& height) : width(width), height(height) {
     setProjectionMatrix(scale3D(2 / width, 2 / height, 1));
     setPosition(vec3(position));
     Camera2D::updateCameraMatrix();
@@ -83,7 +83,7 @@ void Camera2D::fill(const XMLNode& node) {
 
 //Camera3D
 
-Camera3D::Camera3D(const vec3& pos, const vec3& rot, const float& fov, const float& ratio) : rotation(rot), fov(fov),
+Camera3D::Camera3D(const vec3& pos, const vec3& rot, const double& fov, const double& ratio) : rotation(rot), fov(fov),
                                                                                              ratio(ratio) {
     setProjectionMatrix(scale3D(1,1,0.01f) * perspective(1, 2, fov, ratio));
     setPosition(pos);
@@ -100,28 +100,28 @@ void Camera3D::setRotation(const vec3& rot) {
     updateCameraMatrix();
 }
 
-void Camera3D::setPitch(const float& pitch) {
+void Camera3D::setPitch(const double& pitch) {
     rotation.x = pitch;
     if (rotation.x > 2 * PI)
         rotation.x -= 2 * PI;
     updateCameraMatrix();
 }
 
-void Camera3D::setYaw(const float& yaw) {
+void Camera3D::setYaw(const double& yaw) {
     rotation.y = yaw;
     if (rotation.y > 2 * PI)
         rotation.y -= 2 * PI;
     updateCameraMatrix();
 }
 
-void Camera3D::setRoll(const float& roll) {
+void Camera3D::setRoll(const double& roll) {
     rotation.z = roll;
     if (rotation.z > 2 * PI)
         rotation.z -= 2 * PI;
     updateCameraMatrix();
 }
 
-void Camera3D::setRotation(const float& pitch, const float& yaw, const float& roll) {
+void Camera3D::setRotation(const double& pitch, const double& yaw, const double& roll) {
     rotation.x = pitch;
     rotation.y = yaw;
     rotation.z = roll;
@@ -132,7 +132,7 @@ void Camera3D::setDirection(const vec3& direction) {
     setDirection(direction, 0);
 }
 
-void Camera3D::setDirection(const vec3& direction, const float& roll) {
+void Camera3D::setDirection(const vec3& direction, const double& roll) {
     vec3 d = direction.normalize();
     setRotation(asin(-d.y), atan2(d.x, d.z), roll);
     updateCameraMatrix();
@@ -147,12 +147,12 @@ vec3 Camera3D::getDirection() const {
     return (rotateXZ3D(-rotation.y) * rotateYZ3D(-rotation.x) * rotateXY3D(-rotation.z) * vec4({0, 0, 1, 0})).xyz();
 }
 
-void Camera3D::setFoV(const float& f) {
+void Camera3D::setFoV(const double& f) {
     fov = f;
     setProjectionMatrix(perspective(0.01f, 100, fov, ratio));
 }
 
-void Camera3D::setRatio(const float& rat) {
+void Camera3D::setRatio(const double& rat) {
     ratio = rat;
     setProjectionMatrix(perspective(0.01f, 100, fov, ratio));
 }
