+++ zoe/src/zoe/display/Physics/AxisAlignedBox.h
@@ -12,7 +12,7 @@ class AxisAlignedBox {
 public:
     AxisAlignedBox() = default;
     AxisAlignedBox(const vec3& lower, const vec3& higher);
-    AxisAlignedBox(float x1, float y1, float z1, float x2, float y2, float z2);
+    AxisAlignedBox(double x1, double y1, double z1, double x2, double y2, double z2);
 
     [[nodiscard]] inline const vec3& getLowerCorner() const { return lower; }
     [[nodiscard]] inline const vec3& getHigherCorner() const { return higher; }
